+++ packages/amplify-graphql-api-construct-tests/src/__tests__/backends/amplify-table/canary/app.ts
@@ -1,0 +1,30 @@
+#!/usr/bin/env node
+import 'source-map-support/register';
+import { App, Stack, Duration } from 'aws-cdk-lib';
+// @ts-ignore
+import { AmplifyGraphqlApi, AmplifyGraphqlDefinition } from '@aws-amplify/graphql-api-construct';
+
+const packageJson = require('../package.json');
+
+const app = new App();
+const stack = new Stack(app, packageJson.name.replace(/_/g, '-'), {
+  env: { region: process.env.CLI_REGION || 'us-west-2' },
+});
+
+new AmplifyGraphqlApi(stack, 'GraphqlApi', {
+  apiName: 'MyGraphQLApi',
+  definition: AmplifyGraphqlDefinition.fromString(
+    /* GraphQL */ `
+      type Todo @model @auth(rules: [{ allow: public }]) {
+        description: String!
+      }
+    `,
+    {
+      dbType: 'DYNAMODB',
+      provisionStrategy: 'AMPLIFY_TABLE',
+    },
+  ),
+  authorizationModes: {
+    apiKeyConfig: { expires: Duration.days(7) },
+  },
+});