+++ packages/amplify-e2e-tests/CHANGELOG.md
@@ -3,6 +3,18 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [3.21.1](https://github.com/aws-amplify/amplify-category-api/compare/amplify-category-api-e2e-tests@3.21.0...amplify-category-api-e2e-tests@3.21.1) (2024-04-16)
+
+### Bug Fixes
+
+- **api:** protect sql relational fields when using owner rule ([#2463](https://github.com/aws-amplify/amplify-category-api/issues/2463)) ([53cabe2](https://github.com/aws-amplify/amplify-category-api/commit/53cabe2316a32c5cc6e2f9522d82aa92fd5c0689))
+
+# [3.21.0](https://github.com/aws-amplify/amplify-category-api/compare/amplify-category-api-e2e-tests@3.20.3...amplify-category-api-e2e-tests@3.21.0) (2024-04-11)
+
+### Features
+
+- Fetch SNS topic ARN from SQL manifest ([#2345](https://github.com/aws-amplify/amplify-category-api/issues/2345)) ([fca256e](https://github.com/aws-amplify/amplify-category-api/commit/fca256e7cabf5af838b28b26c4ae0c3c8b1583eb))
+
 ## [3.20.3](https://github.com/aws-amplify/amplify-category-api/compare/amplify-category-api-e2e-tests@3.20.2...amplify-category-api-e2e-tests@3.20.3) (2024-03-28)
 
 **Note:** Version bump only for package amplify-category-api-e2e-tests