+++ packages/amplify-graphql-schema-test-library/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [2.2.26](https://github.com/aws-amplify/amplify-category-api/compare/@aws-amplify/graphql-schema-test-library@2.2.25...@aws-amplify/graphql-schema-test-library@2.2.26) (2024-04-16)
+
+**Note:** Version bump only for package @aws-amplify/graphql-schema-test-library
+
+## [2.2.25](https://github.com/aws-amplify/amplify-category-api/compare/@aws-amplify/graphql-schema-test-library@2.2.24...@aws-amplify/graphql-schema-test-library@2.2.25) (2024-04-11)
+
+**Note:** Version bump only for package @aws-amplify/graphql-schema-test-library
+
 ## [2.2.24](https://github.com/aws-amplify/amplify-category-api/compare/@aws-amplify/graphql-schema-test-library@2.2.23...@aws-amplify/graphql-schema-test-library@2.2.24) (2024-03-28)
 
 **Note:** Version bump only for package @aws-amplify/graphql-schema-test-library